+++ apps/api/src/app/endpoints/data-providers/ghostfolio/ghostfolio.service.ts
@@ -1,3 +1,4 @@
+import { SymbolService } from '@ghostfolio/api/app/symbol/symbol.service';
 import { ConfigurationService } from '@ghostfolio/api/services/configuration/configuration.service';
 import { DataProviderService } from '@ghostfolio/api/services/data-provider/data-provider.service';
 import { GhostfolioService as GhostfolioDataProviderService } from '@ghostfolio/api/services/data-provider/ghostfolio/ghostfolio.service';
@@ -25,6 +26,7 @@ import {
   HistoricalResponse,
   LookupItem,
   LookupResponse,
+  MarketDataOfMarketsResponse,
   QuotesResponse
 } from '@ghostfolio/common/interfaces';
 import { UserWithSettings } from '@ghostfolio/common/types';
@@ -32,6 +34,7 @@ import { UserWithSettings } from '@ghostfolio/common/types';
 import { Injectable, Logger } from '@nestjs/common';
 import { DataSource, SymbolProfile } from '@prisma/client';
 import { Big } from 'big.js';
+import { isEmpty } from 'lodash';
 
 @Injectable()
 export class GhostfolioService {
@@ -42,7 +45,8 @@ export class GhostfolioService {
     private readonly dataProviderService: DataProviderService,
     private readonly fetchService: FetchService,
     private readonly prismaService: PrismaService,
-    private readonly propertyService: PropertyService
+    private readonly propertyService: PropertyService,
+    private readonly symbolService: SymbolService
   ) {}
 
   public async getAssetProfile({ symbol }: GetAssetProfileParams) {
@@ -191,6 +195,33 @@ export class GhostfolioService {
       await Promise.all(promises);
 
       return result;
+    } catch (error) {
+      this.logger.error(error);
+
+      throw error;
+    }
+  }
+
+  public async getMarketDataOfMarkets({
+    includeHistoricalData
+  }: {
+    includeHistoricalData: number;
+  }): Promise<MarketDataOfMarketsResponse> {
+    try {
+      const marketDataOfMarkets =
+        await this.symbolService.getMarketDataOfMarkets({
+          includeHistoricalData
+        });
+
+      for (const symbolItem of Object.values(
+        marketDataOfMarkets.fearAndGreedIndex
+      )) {
+        if (!isEmpty(symbolItem)) {
+          symbolItem.dataSource = DataSource.GHOSTFOLIO;
+        }
+      }
+
+      return marketDataOfMarkets;
     } catch (error) {
       this.logger.error(error);