+++ libs/common/src/lib/interfaces/activities.interface.ts
@@ -5,10 +5,16 @@ import { Order, Tag } from '@prisma/client';
 
 export interface Activity extends Order {
   account?: AccountWithPlatform;
+  assetProfile: EnhancedSymbolProfile;
   error?: ActivityError;
   feeInAssetProfileCurrency: number;
   feeInBaseCurrency: number;
-  SymbolProfile: EnhancedSymbolProfile;
+
+  /**
+   * @deprecated Use `assetProfile` instead
+   */
+  SymbolProfile?: EnhancedSymbolProfile;
+
   tagIds?: string[];
   tags?: Tag[];
   unitPriceInAssetProfileCurrency: number;